Update caesars-cipher.json
Caesar's Cipher challenge for Free Code Camp's JavaScript Algorithms and Data Structures certification. The program checks each character, and rebuilds the string one-by-one. Shifting letters 13 places forward is done by checking if it's in the A-M or N-Z array, and replacing it with the item in the same index as the other one. Non-letter characters are unchanged.